The decision whether to treat an object as a generic Object or as an Aware object is now made at compile time--two different method signatures for intern()--rather than at run time via instanceof.  That gives the programmer the freedom to intern an Aware object >as if it were a generic Object<, if desired, simply by casting it to Object in the method call.
Also added "install" methods (two signatures as above) with which an object can
be forcibly installed in the map, supplanting any previously-interned object
that was equal to it.  This is a weird thing to do and the programmer is
responsible for any references that might still be held to the old object, but
at times it comes in handy.